se_open_data/csv.rb - refactor into classes within CSV namespace, in separate files
So, put "all the stuff" except more generic things like DslProxy into
the SeOpenData::CSV::Schema namespace (which is a class)

Simplify the definition of a schema - I don't actually think it really
needs its own DSL now, just a constructor invocation will do.

Simplify the definition of a conversion between schemas - also, this
can be done without a DSL per-se, just a function call (to define the
main parameters) plus a block (to define the row-wise transforms)
should do.

DslProxy is actually no longer used, but we'll keep it for now.