tools/deploy/cronjob - use sudo on deployment commands
They need to be able to write as other users. Use -n to prevent
password prompts, and -E to preserve the password env vars when
deploying as root.

We need to use the full path to seod to allow it to be whitelisted in
sudoers config.

Something like this (deployed currently using ansible in the
technology-and-infrastructure repo):

    # Allow seopendata user to run seod command as root (for deploying open data)
    seopendata ALL = NOPASSWD:SETENV: /home/seopendata/working/tools/se_open_data/bin/seod