se_open_data/csv/converter/limesurveycore.rb - add new fields to schema
The new activity-modified vocab terms: A130, A140, A150 and A160. Add
meta-data which reflects the ESSGlobal vocabs activity-modified and
organisational-structure, and add processing to convert the new
activity values. (Lime Survey for Mersey doesn't add new org struct
values).

Note that this BREAKS OXFORD AND NEWBRIDGE, which don't yet have these
new fields. Rock and a hard place. I will work around this in an
upcoming commit which can be reverted simply when (if?) this is
rectified.