se_open_data/cli.rb - don't delete the virtuoso data dir before triplestore upload
Can't do this because the all.rdf is written there by the `generate`
step.

Need to adjust that to write elsewhere, which requires a bit of a code
review to discover what else relies on its current location, and I
don't have time for that right now.  This is just a quick fix.