ica/converter - hackily munge in Country ID and Territory ID fields #1
We don't want to interfere with the geolocation (because that requires
a lot of work checking and verifying when they inevitably change). So
we will hack in the extra fields after the geolocation, even though we
know that it's wrong in certain cases (i.e. the Supranational members,
and any with an address not in their territory, like ICA#1410).

This is a preliminary stage to allow us to get to a working ICA map
before the ICA adjust their data in the way we are recommending.

This commit 1 of 2: just put the country name verbatim into country_id
fo now (so geocoding works as before, just on a differently named
field), compute territory_id.  Note, this will break other projects
because the standard.csv schema has changed.