ica/converter - hackily munge in Country ID and Territory ID fields #2
commit 2 of 2: add a final hack step which fills in the country id
properly, now that geolocation has been done.

I've checked and the geolocation stays as it was, which is the main
aim of this change. The geolcation code needs to be refactored at some
point, but at least this means we can now generate data for ICA.