se_open_data/* - add a country_id property for initiatives in RDF generation
Rename country_uri to country_id, because the field name seems to be
inserted literally into the RDF, just camelCased. territory_id added
but commented out, because I can't yet work out how to define a
relation to a field which can be one of three vocabs (country, region,
super-region).