coops-uk/converter - adapt to convert Organisations_2023-05-15.csv
This data has a different schema. It needs a new schema.yml file, and
a mapping tabel to re-generate the "SIC Section" field from the SIC
Code, as the former is absent but can mostly be deduced from the SIC
Code->SIC section mapping in the previouos data. However, some mapping
entries needed to be added with a bit of guesswork.