monitor-2022/* - add a converter plus configs
These correct and convert the monitor 2022 data.

Geocoding doesn't work with the country field appended to the address,
there are cases where the address and this country don't combine (as
the country is not the one the address is in!). And this makes the
geocoder throw an exception. Therefore the country field is being
mapped to territory_id.