dotcoop/converter -  add experimental code to geocode addresses
This includes the 3rd party `geocoder` gem, some shim code for
convenience and to allow the details to be tweaked as necessary, a new
cache storage class (actually two), to work with geocoder's
requirements. The cache classes use an in-memory hash, populated on
start and saved on termination. One uses a giant JSON file as before,
the latter uses a directory of files, which seems less
unweildy. Neither are compatible with earlier hash data, because
geocoder wants to store pure key/value - i.e. values are strings not
JSON.

This code will all be moved to the se-open-data gem shortly.