tools/deploy/cronjob - add post_success step
And avoid invoking newer seod commands on older versions.

(post_success is the only real new command we need to be careful of;
but we may remove triplestore later in favour of a deploy step which
includes it; this may be true of generate and convert also. But these
are just anticipations.)