owned-by-oxford/Gemfile - bump se-open-data to 2.2.4
For Cli.fetch with support for custom headers. We need this to use the
new API authorisation.

Plus there are a bunch of other improvements we can use, notably with
geocoding.

We need to include a workaround for stringio - uncertain why, could be
we need to start using Ruby 3, but it works for now.